config: reorganize folder
