+++ src/auction/index.ts
@@ -1,0 +1,1 @@
+export { default as SellForm } from './SellForm';
